requirements.txt
streamlit
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
tiktoken